Refactor: 캐러셀 개선
- 더미데이터 변경
- 캐러셀 컴포넌트화 개선
- 파일, 이름 변경
