+++ src/pages/home/HomePage.tsx
@@ -1,34 +1,10 @@
-import Slider from "react-slick";
+import React from "react";
+import HeroCarousel from "./HeroCarousel";
 import heroData from "./dummydata.json";
-import HeroSection from "./HeroSection";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
 
-const settings = {
-  dot: true,
-  infinite: true,
-  speed: 500,
-  slidesToShow: 1,
-  slidesToScroll: 1,
-  autoplay: true,
-  vertical: true,
-  autoplaySpeed: 2500
-};
-
-const HomePage = () => {
+const HomePage: React.FC = () => {
   return (
-    <Slider {...settings}>
-      {heroData.teams.map((team) => (
-        <HeroSection
-          key={team.id}
-          heroImage={team.heroImage}
-          heroTitle={team.heroTitle}
-          heroText={team.heroText}
-          teamLogo={team.teamLogo}
-          altText={`${team.name} Logo`}
-        />
-      ))}
-    </Slider>
+    <HeroCarousel teams={heroData.teams} />
   );
 };
